Time: 63 ms (25.02%), Space: 14 MB (10.26%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.py
@@ -18,4 +18,4 @@ class Solution(object):
         for i in salary:
             result+=i
             count+=1
-        return float(result/count)
+        return (result/count)
